restructure project - make lib crate - separate file for separate concerns
